refactor: extract metadata + journey extraction from content-fetcher
Phase 2 of the content-fetcher refactor. Move 11 metadata + journey
extraction symbols from src/docs-retrieval/content-fetcher.ts to a new
file src/docs-retrieval/metadata-extractor.ts (Pattern A pure utility
+ Pattern J contract surface for the index.json fetch and the
asymmetric urlsMatch normalization).

Symbols moved:
- extractMetadata (only public consumer, called by fetchContent)
- extractTitleFromJson, extractTitleFromHtml
- extractLearningJourneyMetadata, extractSingleDocMetadata
- extractJourneySummary, extractDocSummary
- getLearningJourneyBaseUrl
- fetchLearningJourneyMetadataFromJson
- findCurrentMilestoneFromUrl
- urlsMatch

The cluster is self-contained: no module-private helpers stay behind in
content-fetcher; the new module needs only types from
../types/content.types and the global fetch API. content-fetcher.ts
now imports `extractMetadata` from `./metadata-extractor` and drops the
now-unused ContentMetadata + SingleDocMetadata type imports.

Critical invariants preserved verbatim (each guarded by post-test
assertions and the Phase 2 hard-gates in PLAN.md):

  * urlsMatch normalization (trailing slash + lowercase) — asymmetric
    vs. fetchPackageContent's strict m.url === contentUrl match.
  * fetchLearningJourneyMetadataFromJson does NOT enforce trust or
    HTTPS on baseUrl (INVESTIGATION §6 invariant 1) — verified by
    pre-test assertion 7 (http:// baseUrl reaches fetch and returns
    milestones).
  * Title regex precedence: <title> → <h1> → og:title → 'Documentation'.
  * findCurrentMilestoneFromUrl /unstyled.html and /content.json
    suffix strip + legacy /milestone-N regex fallback.
  * getLearningJourneyBaseUrl regex precedence:
      learning-journeys → learning-paths → tutorials → strip-milestone.
  * Milestone renumbering: grafana.skip filter + sequential index+1.

No public API changes; no consumer migrations. None of the 11 symbols
were on the docs-retrieval barrel or imported externally — Grep
confirmed zero external consumers.

Pre-test imports flipped from './content-fetcher' to
'./metadata-extractor' as an ownership-path update (wiki Pattern A/J
permitted exception). All 24 pre-extraction assertions still pass
against the new module. Full docs-retrieval test slice passes
(140 tests across 9 suites). typecheck clean.

Refs: .refactor/PLAN.md Phase 2, INVESTIGATION §6 invariants 1 + 5,
DR-04 (no new barrel exports), DR-05 (FetchError stays in raw-fetch
in Phase 4 — no impact here), DR-07 (temporary test exports → permanent).
